float to short and 2 new const strings

+++ Runtime/Global/GameEvents.cs
@@ -3,8 +3,8 @@ public static class GameEvents
     public static System.Action<bool> OnTogglePause;
     public static System.Action<bool> OnToggleTab;
     public static System.Action OnFOVChanged;
-    public static System.Action<float, string> OnObjectPaintingCompleted;
-    public static System.Action<float> OnCurrentMoneyChanged;
+    public static System.Action<short, string> OnObjectPaintingCompleted;
+    public static System.Action<short> OnCurrentMoneyChanged;
 
     public static void TriggerTogglePause(bool isPaused)
     {
@@ -21,12 +21,12 @@ public static class GameEvents
         OnToggleTab?.Invoke(isTabOpen);
     }
 
-    public static void TriggerObjectPaintingCompleted(float cashReward, string rewardStatID)
+    public static void TriggerObjectPaintingCompleted(short cashReward, string rewardStatID)
     {
         OnObjectPaintingCompleted?.Invoke(cashReward, rewardStatID);
     }
 
-    public static void TriggerCurrentMoneyChanged(float newAmount)
+    public static void TriggerCurrentMoneyChanged(short newAmount)
     {
         OnCurrentMoneyChanged?.Invoke(newAmount);
     }
